requirements.txt
discord.py
requests
```
pip install -r requirements.txt
```


